playlist page ,like feature integrated
