+++ package.json
@@ -50,6 +50,7 @@
     "react-dom": "^18.2.0",
     "react-hook-form": "^7.45.4",
     "react-i18next": "^14.1.2",
+    "react-icons": "^5.2.1",
     "react-redux": "^8.1.1",
     "react-router-dom": "^6.14.1",
     "react-scripts": "5.0.1",
