+++ src/components/shared/Header/index.tsx
@@ -27,9 +27,9 @@ const Header: React.FC = () => {
       <HStack.fullFlex style={{ height: '100px', alignItems: 'center' }}>
         <HStack.rowg1 style={{ alignItems: 'center' }}>
           <StyledLogo size={'40px'} strokeWidth={1.5}></StyledLogo>
-          <Heading3 style={{ marginTop: '0.5rem' }}>Tarana 2.0</Heading3>
+          <Heading3 style={{ marginTop: '0.5 rem' }}>Tarana 2.0</Heading3>
         </HStack.rowg1>
-        <HStack.rowg1>
+        <HStack.rowg1 >
           <ThemeToggler />
           <AuthStatus header />
           <NewPageBtn/>
