+++ src/containers/PlaylistContainer/index.tsx
@@ -1,0 +1,35 @@
+import React from 'react';
+import { useSelector } from 'react-redux';
+import { RootState } from '@/store/store';
+import SongCard from '@/components/custom/SongCard';
+
+const LikedSongs: React.FC = () => {
+  const likedSongs = useSelector((state: RootState) => state.songs.likedSongs);
+
+  const handleTogglePlay = (previewUrl: string) => {
+    // Implement play/pause functionality if needed
+    console.log(`Toggled play for song with preview URL: ${previewUrl}`);
+  };
+
+  return (
+    <div>
+      <h1>Liked Songs</h1>
+      {likedSongs.length === 0 ? (
+        <p>No liked songs</p>
+      ) : (
+        <ul>
+          {likedSongs.map((song) => (
+            <li key={song.previewUrl}>
+              <SongCard
+                song={song}
+                onTogglePlay={() => handleTogglePlay(song.previewUrl)}
+              />
+            </li>
+          ))}
+        </ul>
+      )}
+    </div>
+  );
+};
+
+export default LikedSongs;
