+++ src/views/PlayerView/index.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import { Col, Container, Row } from 'react-bootstrap';
 import AudioPlayer from '@/components/custom/AudioPlayer';
-import LikedSongsPage from '@/components/shared/UserPlay';
+
 interface PlayerViewProps {
   isPlaying: boolean;
   activeSong: Song | null;
